Move classreg.h to classreg.hpp
