+++ src/ivoc/symchoos.cpp
@@ -58,7 +58,7 @@
 #include "ivoc.h"
 #endif /* HAVE_IV */
 
-#include "classreg.h"
+#include "classreg.hpp"
 #include "gui-redirect.h"
 
 #if HAVE_IV
