+++ src/nrniv/nrnpy.cpp
@@ -8,7 +8,7 @@
 #include <stdlib.h>
 #include <InterViews/resource.h>
 #include "nrnoc2iv.h"
-#include "classreg.h"
+#include "classreg.hpp"
 #include "nonvintblock.h"
 #include "nrnmpi.h"
