+++ src/nrniv/nrnste.cpp
@@ -4,7 +4,7 @@
 #include <string.h>
 #include <nrnoc2iv.h>
 #include <nrniv_mf.h>
-#include <classreg.h>
+#include <classreg.hpp>
 #include <objcmd.h>
 #include <nrnste.h>
 #include <netcon.h>
