+++ src/nrniv/shapeplt.cpp
@@ -1,5 +1,5 @@
 #include <../../nrnconf.h>
-#include "classreg.h"
+#include "classreg.hpp"
 #include "gui-redirect.h"
 #include "ocnotify.h"
