+++ src/nrnpython/nrnpy_p2h.cpp
@@ -4,7 +4,7 @@
 
 #include <InterViews/resource.h>
 #include <nrnoc2iv.h>
-#include <classreg.h>
+#include <classreg.hpp>
 #include "nrnpython.h"
 #include "hoccontext.h"
 #include "nrnpy.h"
