+++ test/unit_tests/oc/hoc_interpreter.cpp
@@ -2,10 +2,9 @@
 #include <catch2/matchers/catch_matchers_string.hpp>
 #include <catch2/catch_test_macros.hpp>
 
-#include "classreg.h"
+#include "classreg.hpp"
 #include "code.h"
 #include "hocdec.h"
-#include "classreg.h"
 #include "ocfunc.h"
 
 #include <sstream>
